feat(dmsg)+tinygo: extract QUIC behind an interface; fix two latent QUIC stream bugs
TinyGo Phase 2: decouple quic-go from the dmsg session core so a TinyGo (IoT)
dmsg client compiles. quic-go does not build under TinyGo (needs
crypto/tls.QUICEncryptionLevel), and it was woven into six session-core files as
*quic.Conn / *quic.Stream struct fields. Now the core refers to QUIC only via
the quicConn / quicStream interfaces (quic_iface.go); the concrete adapters,
session constructors, client dial and server listener live in quic_native.go
(//go:build !tinygo), with TinyGo stubs in quic_stub.go that fall back to TCP/WS.
wt.go (WebTransport, also quic-go) gets //go:build !tinygo. Result: pkg/dmsg/dmsg
no longer pulls quic-go/webtransport-go under the tinygo tag.

Writing the first dmsg-over-QUIC round-trip test (quic_test.go — a QUIC-only
server, two clients, a bridged A→B stream) surfaced TWO pre-existing bugs in the
dmsg-over-QUIC stream path that no test had ever exercised:

  - server forwardRequest() opened the destination stream with a smux/else-yamux
    branch — for a QUIC destination it dereferenced a nil *yamux.Session and
    panicked serveStream. Added the QUIC case (OpenStreamSync).
  - the client accept loop never treated a QUIC AcceptStream error as terminal
    (no quic case in the session-closed switch), so it span forever on a closed
    QUIC session and hung ClientSession.Close(). quic-go's AcceptStream only
    errors on a done connection, so any error there is now terminal.

Supporting TinyGo proofing (logrus/json compile under TinyGo 0.41, unlike the
#2836-era assumption): pkg/logging now builds its real logrus implementation on
all targets (dropped the //go:build !tinygo tags + the no-op logging_tinygo.go
stub, which couldn't satisfy logrus.FieldLogger for the full client anyway); the
VictoriaMetrics-backed Metrics impl is tagged !tinygo (server-only; the client
uses the no-op Empty).

Regular go build ./... + the full pkg/dmsg/dmsg suite (incl. the new QUIC
round-trip) stay green; the QUIC extraction is verified quic-free under
`go list -tags tinygo`. The remaining TinyGo blocker is pkg/netutil
(interface-enumeration), tracked as Phase 3 in docs/design/tinygo-dmsg-client.md.
