+++ pkg/logging/logging.go
@@ -1,9 +1,5 @@
-//go:build !tinygo
-
 /*
 Package logging provides application logging utilities
-
-Tagged off WASM — see logging_js.go for the build-tag rationale.
 */
 package logging
